refactoring out the idea of hiding components that are on a different tab
