+++ src/connectionconfig/connectionDialogWebViewController.ts
@@ -29,7 +29,8 @@ export class ConnectionDialogWebViewController extends ReactWebViewPanelControll
 				connectionProfile: {} as IConnectionDialogProfile,
 				recentConnections: [],
 				selectedFormTab: FormTabType.Parameters,
-				formTabs: [],
+				connectionFormComponents: [],
+				connectionStringComponents: [],
 				connectionStatus: ApiStatus.NotStarted,
 				formError: ''
 			},
@@ -50,7 +51,10 @@ export class ConnectionDialogWebViewController extends ReactWebViewPanelControll
 		} else {
 			await this.loadEmptyConnection();
 		}
-		this.state.formTabs = await this.generateFormTabs();
+
+		this.state.connectionFormComponents = await this.generateConnectionFormComponents();
+		this.state.connectionStringComponents = await this.generateConnectionStringComponents();
+
 		await this.updateItemVisibility();
 		this.state = this.state;
 	}
@@ -109,23 +113,7 @@ export class ConnectionDialogWebViewController extends ReactWebViewPanelControll
 	private async updateItemVisibility() {
 		const selectedTab = this.state.selectedFormTab;
 		let hiddenProperties: (keyof IConnectionDialogProfile)[] = [];
-		if (selectedTab === FormTabType.ConnectionString) {
-			hiddenProperties = [
-				'server',
-				'authenticationType',
-				'user',
-				'password',
-				'savePassword',
-				'accountId',
-				'tenantId',
-				'database',
-				'trustServerCertificate',
-				'encrypt'
-			];
-		} else {
-			hiddenProperties = [
-				'connectionString'
-			];
+		if (selectedTab === FormTabType.Parameters) {
 			if (this.state.connectionProfile.authenticationType !== AuthenticationType.SqlLogin) {
 				hiddenProperties.push('user', 'password', 'savePassword');
 			}
@@ -138,19 +126,24 @@ export class ConnectionDialogWebViewController extends ReactWebViewPanelControll
 				if (tenants.length === 1) {
 					hiddenProperties.push('tenantId');
 				}
-
 			}
 		}
 
-		for (const tab of this.state.formTabs) {
-			for (const component of tab.components) {
+		for (const component of this.state.connectionFormComponents) {
 				component.hidden = hiddenProperties.includes(component.propertyName);
-			}
 		}
 	}
 
+	private getActiveFormComponents(): FormComponent[] {
+		if (this.state.selectedFormTab === FormTabType.Parameters) {
+			return this.state.connectionFormComponents;
+		}
+		return this.state.connectionStringComponents;
+	}
+
 	private getFormComponent(propertyName: keyof IConnectionDialogProfile): FormComponent | undefined {
-		return this.state.formTabs.flatMap(t => t.components).find(c => c.propertyName === propertyName);
+
+		return this.getActiveFormComponents().find(c => c.propertyName === propertyName);
 	}
 
 	private async getAccounts(): Promise<FormComponentOptions[]> {
@@ -181,194 +174,187 @@ export class ConnectionDialogWebViewController extends ReactWebViewPanelControll
 		});
 	}
 
-	private async generateFormTabs(): Promise<FormTab[]> {
-		const result: FormTab[] = [
+	private async generateConnectionFormComponents(): Promise<FormComponent[]> {
+		return [
 			{
-				tab: FormTabType.Parameters,
-				components: [
-					{
-						type: FormComponentType.Input,
-						propertyName: 'server',
-						label: 'Server',
-						required: true,
-						validate: (value: string) => {
-							if (this.state.selectedFormTab === FormTabType.Parameters && !value) {
-								return {
-									isValid: false,
-									validationMessage: 'Server is required'
-								};
-							}
-							return {
-								isValid: true,
-								validationMessage: ''
-							};
-						}
-					},
-					{
-						type: FormComponentType.Dropdown,
-						propertyName: 'authenticationType',
-						label: 'Authentication Type',
-						required: true,
-						options: [
-							{
-								displayName: 'SQL Login',
-								value: AuthenticationType.SqlLogin
-							},
-							{
-								displayName: 'Windows Authentication',
-								value: AuthenticationType.Integrated
-							},
-							{
-								displayName: 'Microsoft Entra MFA',
-								value: AuthenticationType.AzureMFA
-							}
-						],
-					},
-					{
-						// Hidden if connection string is set or if the authentication type is not SQL Login
-						propertyName: 'user',
-						label: 'User Name',
-						type: FormComponentType.Input,
-						required: true,
-						validate: (value: string) => {
-							if (this.state.connectionProfile.authenticationType === AuthenticationType.SqlLogin && !value) {
-								return {
-									isValid: false,
-									validationMessage: 'User name is required'
-								};
-							}
-							return {
-								isValid: true,
-								validationMessage: ''
-							};
-						}
-					},
-					{
-						propertyName: 'password',
-						label: 'Password',
-						required: false,
-						type: FormComponentType.Password,
-					},
-					{
-						propertyName: 'savePassword',
-						label: 'Save Password',
-						required: false,
-						type: FormComponentType.Checkbox,
-					},
-					{
-						propertyName: 'accountId',
-						label: 'Azure Account',
-						required: true,
-						type: FormComponentType.Dropdown,
-						options: await this.getAccounts(),
-						placeholder: 'Select an account',
-						actionButtons: await this.getAzureActionButtons(),
-						validate: (value: string) => {
-							if (this.state.connectionProfile.authenticationType === AuthenticationType.AzureMFA && !value) {
-								return {
-									isValid: false,
-									validationMessage: 'Azure Account is required'
-								};
-							}
-							return {
-								isValid: true,
-								validationMessage: ''
-							};
-						},
-					},
-					{
-						propertyName: 'tenantId',
-						label: 'Tenant ID',
-						required: true,
-						type: FormComponentType.Dropdown,
-						options: [],
-						hidden: true,
-						placeholder: 'Select a tenant',
-						validate: (value: string) => {
-							if (this.state.connectionProfile.authenticationType === AuthenticationType.AzureMFA && !value) {
-								return {
-									isValid: false,
-									validationMessage: 'Tenant ID is required'
-								};
-							}
-							return {
-								isValid: true,
-								validationMessage: ''
-							};
-						}
-					},
-					{
-						propertyName: 'database',
-						label: 'Database',
-						required: false,
-						type: FormComponentType.Input,
-					},
-					{
-						propertyName: 'trustServerCertificate',
-						label: 'Trust Server Certificate',
-						required: false,
-						type: FormComponentType.Checkbox,
-					},
-					{
-						propertyName: 'encrypt',
-						label: 'Encrypt Connection',
-						required: false,
-						type: FormComponentType.Dropdown,
-						options: [
-							{
-								displayName: 'Optional',
-								value: 'Optional'
-							},
-							{
-								displayName: 'Mandatory',
-								value: 'Mandatory'
-							},
-							{
-								displayName: 'Strict  (Requires SQL Server 2022 or Azure SQL)',
-								value: 'Strict'
-							}
-						],
-					},
-					{
-						propertyName: 'profileName',
-						label: 'Profile Name',
-						required: false,
-						type: FormComponentType.Input,
+				type: FormComponentType.Input,
+				propertyName: 'server',
+				label: 'Server',
+				required: true,
+				validate: (value: string) => {
+					if (this.state.selectedFormTab === FormTabType.Parameters && !value) {
+						return {
+							isValid: false,
+							validationMessage: 'Server is required'
+						};
 					}
-				]
+					return {
+						isValid: true,
+						validationMessage: ''
+					};
+				}
 			},
 			{
-				tab: FormTabType.ConnectionString,
-				components: [
+				type: FormComponentType.Dropdown,
+				propertyName: 'authenticationType',
+				label: 'Authentication Type',
+				required: true,
+				options: [
 					{
-						type: FormComponentType.TextArea,
-						propertyName: 'connectionString',
-						label: 'Connection String',
-						required: true,
-						validate: (value: string) => {
-							if (this.state.selectedFormTab === FormTabType.ConnectionString && !value) {
-								return {
-									isValid: false,
-									validationMessage: 'Connection string is required'
-								};
-							}
-							return {
-								isValid: true,
-								validationMessage: ''
-							};
-						}
+						displayName: 'SQL Login',
+						value: AuthenticationType.SqlLogin
 					},
 					{
-						propertyName: 'profileName',
-						label: 'Profile Name',
-						required: false,
-						type: FormComponentType.Input,
+						displayName: 'Windows Authentication',
+						value: AuthenticationType.Integrated
+					},
+					{
+						displayName: 'Microsoft Entra MFA',
+						value: AuthenticationType.AzureMFA
 					}
-				]
+				],
+			},
+			{
+				// Hidden if connection string is set or if the authentication type is not SQL Login
+				propertyName: 'user',
+				label: 'User Name',
+				type: FormComponentType.Input,
+				required: true,
+				validate: (value: string) => {
+					if (this.state.connectionProfile.authenticationType === AuthenticationType.SqlLogin && !value) {
+						return {
+							isValid: false,
+							validationMessage: 'User name is required'
+						};
+					}
+					return {
+						isValid: true,
+						validationMessage: ''
+					};
+				}
+			},
+			{
+				propertyName: 'password',
+				label: 'Password',
+				required: false,
+				type: FormComponentType.Password,
+			},
+			{
+				propertyName: 'savePassword',
+				label: 'Save Password',
+				required: false,
+				type: FormComponentType.Checkbox,
+			},
+			{
+				propertyName: 'accountId',
+				label: 'Azure Account',
+				required: true,
+				type: FormComponentType.Dropdown,
+				options: await this.getAccounts(),
+				placeholder: 'Select an account',
+				actionButtons: await this.getAzureActionButtons(),
+				validate: (value: string) => {
+					if (this.state.connectionProfile.authenticationType === AuthenticationType.AzureMFA && !value) {
+						return {
+							isValid: false,
+							validationMessage: 'Azure Account is required'
+						};
+					}
+					return {
+						isValid: true,
+						validationMessage: ''
+					};
+				},
+			},
+			{
+				propertyName: 'tenantId',
+				label: 'Tenant ID',
+				required: true,
+				type: FormComponentType.Dropdown,
+				options: [],
+				hidden: true,
+				placeholder: 'Select a tenant',
+				validate: (value: string) => {
+					if (this.state.connectionProfile.authenticationType === AuthenticationType.AzureMFA && !value) {
+						return {
+							isValid: false,
+							validationMessage: 'Tenant ID is required'
+						};
+					}
+					return {
+						isValid: true,
+						validationMessage: ''
+					};
+				}
+			},
+			{
+				propertyName: 'database',
+				label: 'Database',
+				required: false,
+				type: FormComponentType.Input,
+			},
+			{
+				propertyName: 'trustServerCertificate',
+				label: 'Trust Server Certificate',
+				required: false,
+				type: FormComponentType.Checkbox,
+			},
+			{
+				propertyName: 'encrypt',
+				label: 'Encrypt Connection',
+				required: false,
+				type: FormComponentType.Dropdown,
+				options: [
+					{
+						displayName: 'Optional',
+						value: 'Optional'
+					},
+					{
+						displayName: 'Mandatory',
+						value: 'Mandatory'
+					},
+					{
+						displayName: 'Strict  (Requires SQL Server 2022 or Azure SQL)',
+						value: 'Strict'
+					}
+				],
+			},
+			{
+				propertyName: 'profileName',
+				label: 'Profile Name',
+				required: false,
+				type: FormComponentType.Input,
 			}
 		];
+	}
 
-		return result;
+	private async generateConnectionStringComponents(): Promise<FormComponent[]> {
+		return [
+			{
+				type: FormComponentType.TextArea,
+				propertyName: 'connectionString',
+				label: 'Connection String',
+				required: true,
+				validate: (value: string) => {
+					if (this.state.selectedFormTab === FormTabType.ConnectionString && !value) {
+						return {
+							isValid: false,
+							validationMessage: 'Connection string is required'
+						};
+					}
+					return {
+						isValid: true,
+						validationMessage: ''
+					};
+				}
+			},
+			{
+				propertyName: 'profileName',
+				label: 'Profile Name',
+				required: false,
+				type: FormComponentType.Input,
+			}
+		];
 	}
 
 	private async validateFormComponents(propertyName?: keyof IConnectionDialogProfile): Promise<number> {
@@ -383,7 +369,7 @@ export class ConnectionDialogWebViewController extends ReactWebViewPanelControll
 			}
 		}
 		else {
-			this.state.formTabs.flatMap(t => t.components.forEach(c => {
+			this.getActiveFormComponents().forEach(c => {
 				if (c.hidden) {
 					c.validation = {
 						isValid: true,
@@ -398,7 +384,7 @@ export class ConnectionDialogWebViewController extends ReactWebViewPanelControll
 						}
 					}
 				}
-			}));
+			});
 		}
 		return errorCount;
 	}
@@ -484,10 +470,8 @@ export class ConnectionDialogWebViewController extends ReactWebViewPanelControll
 
 	private clearFormError() {
 		this.state.formError = '';
-		for (const tab of this.state.formTabs) {
-			for (const component of tab.components) {
-				component.validation = undefined;
-			}
+		for (const component of this.getActiveFormComponents()) {
+			component.validation = undefined;
 		}
 	}
 
@@ -531,7 +515,7 @@ export class ConnectionDialogWebViewController extends ReactWebViewPanelControll
 			this.state.connectionStatus = ApiStatus.Loading;
 			this.state.formError = '';
 			this.state = this.state;
-			const visibleComponents = this.state.formTabs.flatMap(t => t.components.filter(c => !c.hidden).map(c => c.propertyName));
+			const visibleComponents = this.getActiveFormComponents().filter(c => !c.hidden).map(c => c.propertyName);
 			// Set all other fields to undefined
 			Object.keys(this.state.connectionProfile).forEach(key => {
 				if (!visibleComponents.includes(key as keyof IConnectionDialogProfile)) {
